requirements.txt
django>=3.2.3,<3.3
django-cors-headers==3.6.0
django-grappelli==2.15.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-spectacular==0.21.0
lorem==0.1.1
Pillow
python-dotenv==0.19.2
pytz==2021.3